feat(ast): add terminated statement
The terminated statement represents a statement which is optionally
terminated by a semicolon. GML is very, very relaxed about where
semicolons can and cannot appear, so allowing them after essentially any
statement is one of the better options for parsing. Enforcing semicolons
in only certain positions is a job better left to a linter or a semantic
analysis pass.